22_data_quality_monitor/requirements.txt
psycopg2-binary>=2.9
pydantic-settings>=2.0
python-dotenv>=1.0
